requirements.txt
asgiref==3.5.0
cloudinary==1.29.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-uuslug==2.0.0
gunicorn==20.1.0
psycopg2==2.9.3
python-slugify==6.1.1
pytz==2021.3
sqlparse==0.4.2
text-unidecode==1.3
